// src/types/types.ts
export type CryptoCurrency = {
  id: string;
  rank: string;
  symbol: string;
  name: string;
  supply: string;
  maxSupply: string;
  marketCapUsd: string;
  volumeUsd24Hr: string;
  priceUsd: string;
  changePercent24Hr: string;
  vwap24Hr: string;
  explorer: string;
};
// export type CryptoCurrency = {
//   exchange_idstring,
//   website: string;
//   name: string;
//   data_quote_start: string;
//   data_quote_end: string;
//   data_orderbook_start: string;
//   data_orderbook_end: string;
//   data_trade_start: string;
//   data_trade_end: string;
//   data_symbols_count: string;
//   volume_1hrs_usd: string;
//   volume_1day_usd: string;
//   volume_1mth_usd: string;
// };